task3 and task4

+++ labs/lab7/task3.c
@@ -4,11 +4,11 @@
 
 int main() {
     sem_t sem;
-    int max_value = 0;
+    long max_value = 0;
 
     while (1) {
         if (sem_init(&sem, 0, max_value) != 0) {
-            printf("Max counting semaphore value: %d\n", max_value);
+            printf("Max counting semaphore value: %ld\n", max_value);
             break;
         }
         max_value++;
